order accepts float instead of int

+++ backend/api/views.py
@@ -425,7 +425,7 @@ class RazorpayCheckoutAPIView(generics.CreateAPIView):
         try:
             # Create Razorpay Order
             razorpay_order = razorpay_client.order.create({
-                'amount': int(order.total) * 100, 
+                'amount': float(order.total) * 100, 
                 'currency': 'INR',
                 'receipt': order.oid,
                 'payment_capture': '1'
@@ -437,7 +437,7 @@ class RazorpayCheckoutAPIView(generics.CreateAPIView):
             # Prepare checkout data for frontend
             checkout_data = {
                 'key': settings.RAZORPAY_KEY_ID,
-                'amount': int(order.total),
+                'amount': float(order.total),
                 'currency': 'INR',
                 'name': 'Web3Lms',
                 'description': f'Payment for order {order.oid} of {order.total} by {order.full_name} ({order.email})',
